katana-mngr/requirements.txt
Flask==1.0.2

# API.
flask-classful==0.14

PyYAML==3.13

# Application server for both development and production.
gunicorn==19.9.0

multiprocess>=0.70.7

openstacksdk==0.23.0

# PyMongo
pymongo==3.7.2

# Tnglib
git+https://github.com/sonata-nfv/tng-cli